+++ src/styles/components/_authPage.scss
@@ -64,6 +64,7 @@
     font-weight: bold;
     font-size: 14px;
     line-height: 17px;
+    color: #FC842D;
     &:hover{
         background-color: #FC842D;
         color: #FFFFFF;
